0.1.5: Use libreoffice meta-package. Allow -* app-removals
It is safest to install everything from libreoffice,
so make sure the meta-package is installed. That
automatically included the other desired components
except for the legacy hsqldb - which actually is
not yet deprecated yet, so someone is way too
early not including that by default.

Way too many fonts were showing in LibreOffice,
so I discovered that the removal script wasn't
working for any *'d entries. Fixed that.

python-appindicator isn't (currently) available
as a package, so it was blocking everything...